update/ adjust component
