+++ modules/UserCenter/MyGroups/index.tsx
@@ -1,11 +1,12 @@
 import React, { useState } from "react";
+import TabSection from "../TabSection";
+import GroupsList from "./GroupsList";
 import Link from "@/common/components/GeneralLink";
 import Button from "@/common/components/GeneralButton";
 import groupStatus from "@/constants/groupStatus";
-import TabSection from "../TabSection";
 import { groupsData, tabs, tabIdType } from "../date";
 
-export default function MyGroups() {
+export default function MyGroups({}) {
   const [activeTab, setActiveTab] = useState<tabIdType>("upcoming");
 
   const handleCancelApply = () => {};
@@ -34,9 +35,20 @@ export default function MyGroups() {
           </p>
         </div>
         <ul>
+          {/* <GroupsList activeTab={activeTab} /> */}
           {groupsData.map((group) => {
-            const { endTime, status } = group;
-            // if (status === "leader") return;
+            const {
+              groupId,
+              groupName,
+              store,
+              place,
+              totalMemberNum,
+              currentNum,
+              startTime,
+              endTime,
+              status,
+            } = group;
+
             const setStatus = () => {
               const now = new Date();
               const today = now.toISOString();
@@ -49,7 +61,49 @@ export default function MyGroups() {
               (activeTab === "over" && setStatus() === "over") ||
               (activeTab === "upcoming" && setStatus() !== "over");
             if (!isActive) return;
-            return <></>;
+
+            const statusStyle = groupStatus[setStatus()].style;
+            const statusText = groupStatus[setStatus()].text;
+
+            const isPlace = place === "NULL";
+            const location = isPlace ? store.storeName : place;
+            const groupDate = startTime.split("T")[0];
+            const formatStartTime = startTime.split("T")[1].substring(0, 5);
+            const formatEndTime = endTime.split("T")[1].substring(0, 5);
+            const groupTime = `${groupDate} ${formatStartTime} - ${formatEndTime}`;
+            // 寫一個 helper formatTime
+
+            return (
+              <li
+                key={groupId}
+                className="w-full flex justify-between items-center gap-3 md:gap-2 p-2 mb-3 bg-yellow-tint text-center text-sm"
+              >
+                <p className="w-[10%] text-xs">
+                  <span className={`groupStatusDot ${statusStyle}`}>
+                    {statusText}
+                  </span>
+                </p>
+                <p className="w-[20%] truncate text-sm">
+                  <Link href={`/group/${groupId}`}>{groupName}</Link>
+                </p>
+                <p className="w-[20%] truncate">{location}</p>
+                <p className="w-[20%]">{groupTime}</p>
+                <p className="w-[10%]">
+                  {currentNum}/{totalMemberNum}
+                </p>
+                <p className="w-[10%]">
+                  <Button
+                    type="button"
+                    appearance="black"
+                    rounded
+                    onClick={handleCancelApply}
+                  >
+                    取消申請
+                    {/* 這個要改 */}
+                  </Button>
+                </p>
+              </li>
+            );
           })}
         </ul>
       </div>
